Add -m "mode" switch to extsmaild.
This replaces the previous rather confusing "-d" switch. extsmaild with no arguments (as before) or "-m batch" defaults to batch mode delivery. If "-m daemon" is specified, it enters daemon mode. It's possible that in the future variations on these modes will develop, hence why it's better to bite the bullet now and have a single configurable switch rather than a mass of inter-locking switches.