On each send cycle, start on a different message in the spool queue.
Previously, if a message was unsendable (e.g. the remote sendmail refused to
send it because it is too big), it could stop other messages ever being sent.
This patch prevents that. On each cycle, we try starting one entry further
into the spool dir. In other words if we have entries:
   a, b, c
in the spool dir then we will try sending them in this order upon
each cycle (until some are actually sent):
   1: a, b, c
   2: b, c, a
   3: c, a, b

This patch also has a little bit of stub support for future message reporting
to the user. This is currently largely unused.