Add support for sending notifications when message send cycles have been unsuccessful.
This should prove to be rather useful, particularly when on a dodgy network
that tends to go down without you realising.

The only bad thing is that there seem to be umpteen different notification
frameworks, so this will need a fair bit of FAQ-type documenting to get over
platform differences. Oh well.