Backoff exponentially when retrying unsuccessful sends.
Previously extsmail waited a fixed 60 seconds when retrying sends. This can
be a pain if there was a very temporary problem, or if only 1 message out of
many failed (e.g. because it's too big for a particular mail server). This
patch now retries at intervals of 5, 10, 20, 40, 60 seconds (with 60s being
the maximum), so temporary errors don't cause big delays between tries. A
successful send resets the timer back down to 5s.