Track SIGKILLed processes better so that they can never hold up extsmaild.
Normally SIGKILL has immediate effect, unless the process is an
uninterruptable sleep (e.g. it's waiting on a read from a disk with a bad
sector) at which point the kernel may keep the process alive but not send
signals to it. We therefore maintain a linked list of processes which have
been timed out and for which we're waiting for SIGKILL to take effect.