Once we've forked, ensure that child processes don't try processing data.
When we've found a match, we fork, and exec the command the user has specified.
If there are any problems in that process, we sometimes called "goto next" --
i.e. the child process executed as if it was the parent process. The good news
is that this could never cause mail to be lost (at worst it would be sent more
than once), but it could cause a fork bomb of sorts. The correct behaviour is
for the child process to exit() on any error.