If some messages are being sent successfully, try again soon.
The idea here is that if some messages are being sent successfully in a cycle
but there is a failure, we want to retry again soon, so that all potentially
successful messages are sent as quickly as possible. Only when all messages in
the queue are failures do we start to increase the wait time.