Radically simplify moving through the queue.
The fundamental weakness in the previous scheme was that readdir() not only
doesn't define a specific ordering (e.g. alphabetic ordering), but there's not
even a guarantee of deterministic iteration (i.e. it's feasible to get the same
files in a different order).

This commit does several things:

  1. Each time it tries to send all messages in the queue: it does not stop on
     the first failure.

  2. It rips out the idea of "start at a specific point in the queue the next
     time we try it."

  3. If one or more messages couldn't be sent, it starts at a random point in
     the queue next time.

This is much simpler than the previous scheme, and is much more aggressive at
trying to get messages sent.