Ensure stderr is read even on process failure.
Previously we stopped reading a process's stderr as soon as we noticed a
failure: this made debugging failing subprocesses difficult for the
user. The particular problem in this situation was only triggered for
messages that exceeded the initial buffer size.

This commit fundamentally rethinks how we read/write to child processes,
which fixes the above problem and more:

  1. We try as hard as possible to read all stderr output (even if e.g.
     the child hasn't consumed all input).
  2. We don't close the child's stdin when we notice an error, as
     closing stdin is an implicit signal to the child process that it
     can send an email. Instead, we keep the file open until after a)
     we've `kill`ed the child b) waited upon it. This lessens the chance
     of emails being sent twice.
  3. In the (very unlikely, but not impossible) situation that we can no
     longer read a message from the spool_dir, we kill the child
     process.

To try and wrap my head around all of this, I have also added a basic
test suite of (naughty and nice) child processes (including the bug that
triggered this commit). This highlighted differences in how Linux and
OpenBSD handle various unusual situations: the tests now pass on both
platforms.