Cope with child processes that stall while waiting.
Previously it was possible for a child process to block extsmail
arbitrarily long by reading all the input, closing file descriptors, and
then not exiting. This commit (crudely) adds a timeout to `wait`ing: if
exceeded, the child process is killed. A slightly more efficient
solution would be a SIGCHLD handler and the self-pipe trick.