Don't close stderr twice.
Because POLLIN and POLLHUP are not mutually exclusive, we could
potentially close the stderr file descriptor twice. This is an unlikely
scenario (I haven't encountered it and I can't think of a way of writing
a test that would trigger it), but I believe it is possible.

[An aside: POLLOUT and POLLHUP are mutually exclusive. Why are POLLIN
and POLLHUP not mutually exclusive? It makes code with `poll` quite hard
to write!]